1:AI 2:事件系统

+++ WinSocket_sync_block/WinSocket_sync_block/GUGPackage.h
@@ -196,7 +196,7 @@ namespace GUGGAME
 	struct Walk
 	{
 		unsigned int no;
-		short id;
+		int id;
 		short x;
 		short z;
 		char  dir;
@@ -321,7 +321,7 @@ namespace GUGGAME
 	struct DeleteObj
 	{
 		unsigned int no;
-		short id;
+		int id;
 		DeleteObj()
 		{
 			no = 2003;
